--- src/main/java/com/liftoff/ecommerce/Config/CustomUserDetailsService.java
@@ -1,32 +1,0 @@
-package com.liftoff.ecommerce.Config;
-
-import com.liftoff.ecommerce.Models.User;
-import com.liftoff.ecommerce.Repositories.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Service
-public class CustomUserDetailsService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Autowired
-    public CustomUserDetailsService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userRepository.findByEmail(username);
-        if (user == null) {
-            throw new UsernameNotFoundException("User not found with email: " + username);
-        }
-        return new org.springframework.security.core.userdetails.User(user.getEmail(), user.getPwd(), new ArrayList<>());
-    }
-}
