--- src/main/java/com/liftoff/ecommerce/Models/User.java
@@ -1,103 +1,0 @@
-package com.liftoff.ecommerce.Models;
-
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import jakarta.persistence.*;
-import org.hibernate.annotations.GenericGenerator;
-
-import java.util.HashSet;
-import java.util.Set;
-
-@Entity
-public class User {
-
-        @Id
-        @GeneratedValue(strategy= GenerationType.AUTO,generator="native")
-        @GenericGenerator(name = "native",strategy = "native")
-        @Column(name = "customer_id")
-        private int id;
-
-        private String name;
-
-        private String email;
-
-        @Column(name = "mobile_number")
-        private String mobileNumber;
-
-        @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-        private String pwd;
-
-        private String role;
-
-        @Column(name = "create_dt")
-        private String createDt;
-
-        @JsonIgnore
-        @OneToMany(mappedBy="user",fetch=FetchType.EAGER)
-        private Set<Authority> authorities;
-
-        public int getId() {
-            return id;
-        }
-
-        public void setId(int id) {
-            this.id = id;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
-
-        public String getEmail() {
-            return email;
-        }
-
-        public void setEmail(String email) {
-            this.email = email;
-        }
-
-        public String getMobileNumber() {
-            return mobileNumber;
-        }
-
-        public void setMobileNumber(String mobileNumber) {
-            this.mobileNumber = mobileNumber;
-        }
-
-        public String getPwd() {
-            return pwd;
-        }
-
-        public void setPwd(String pwd) {
-            this.pwd = pwd;
-        }
-
-        public String getRole() {
-            return role;
-        }
-
-        public void setRole(String role) {
-            this.role = role;
-        }
-
-        public String getCreateDt() {
-            return createDt;
-        }
-
-        public void setCreateDt(String createDt) {
-            this.createDt = createDt;
-        }
-
-        public Set<Authority> getAuthorities() {
-            return authorities;
-        }
-
-        public void setAuthorities(Set<Authority> authorities) {
-            this.authorities = authorities;
-        }
-    }
