+++ src/main/java/com/liftoff/ecommerce/Repositories/CustomerRepository.java
@@ -1,0 +1,15 @@
+package com.liftoff.ecommerce.Repositories;
+
+
+import com.liftoff.ecommerce.Models.Customer;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface CustomerRepository extends CrudRepository<Customer, Long> {
+
+    List<Customer> findByEmail(String email);
+
+}
