--- src/main/java/com/liftoff/ecommerce/Repositories/UserRepository.java
@@ -1,16 +1,0 @@
-package com.liftoff.ecommerce.Repositories;
-
-
-import com.liftoff.ecommerce.Models.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    User findByEmail(String email);
-
-}
