+++ src/main/java/com/liftoff/ecommerce/filter/JWTTokenGeneratorFilter.java
@@ -1,4 +1,5 @@
 package com.liftoff.ecommerce.filter;
+
 import com.liftoff.ecommerce.constants.SecurityConstants;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
@@ -34,7 +35,6 @@ public class JWTTokenGeneratorFilter extends OncePerRequestFilter {
                     .setExpiration(new Date((new Date()).getTime() + 30000000))
                     .signWith(key).compact();
             response.setHeader(SecurityConstants.JWT_HEADER, jwt);
-            System.out.println(jwt);
         }
 
         filterChain.doFilter(request, response);
