+++ src/main/java/com/liftoff/ecommerce/filter/RequestValidationBeforeFilter.java
@@ -1,7 +1,6 @@
 package com.liftoff.ecommerce.filter;
 
 import jakarta.servlet.*;
-import jakarta.servlet.Filter;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.security.authentication.BadCredentialsException;
